Rename plugin to @djs-core/my-super-plugin
Renamed the plugin directory and package from 'my-super-djs-plugin' to '@djs-core/my-super-plugin' and updated import paths accordingly in djsconfig.ts. This aligns the naming with scoped package conventions.
