+++ examples/my-super-plugin/index.ts
