+++ examples/my-super-plugin/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-super-djs-plugin",
+  "name": "@djs-core/my-super-plugin",
   "version": "0.1.0",
   "type": "module",
   "main": "index.ts",
